Added a traveserse graph -> get paths class
We get every possible paths starting from an Ontology and avoiding
cycles. We are using depth first instead of breath because we don´t
need the mod optimal path(or the shorthest), but the ones who traverse
most of the objects to have the whole picture in case someone is not
giving attention to our restrictions and conditions in the ontologies.
@TODO. cleanup and remove some unused properties in the class.